Development 1、新增项目 2、增加日志 3、修改注释 4、派遣管理 5、系统设置

+++ src/main/java/com/sw/basis/utils/vo/IdVO.java
@@ -11,17 +11,17 @@ import javax.validation.constraints.NotNull;
 @Data
 public class IdVO {
 
-    public static IdVO newIdVO(Long id){
+    public static IdVO newIdVO(Integer id){
         return new IdVO(id);
     }
 
     @NotNull
-    private Long id;
+    private Integer id;
 
     public IdVO() {
     }
 
-    public IdVO(@NotNull Long id) {
+    public IdVO(@NotNull Integer id) {
         this.id = id;
     }
 }
